requirements.txt
python-jsonrpc
jsonrpclib-pelix
pymongo
pika
redis
beautifulsoup4
requests
scikit-learn
python-dateutil
numpy
pytest-shutil
pandas
tensorflow
